Socket and server now run on the same port
