+++ client/src/pages/LiveChatDoctor.js
@@ -7,7 +7,7 @@ import Axios from 'axios';
 import io from 'socket.io-client';
 
 
-const socket = io.connect("http://localhost:5069");
+const socket = io.connect("http://localhost:8080");
 
 
 function LiveChat() {
